+++ src/evcxr.rs
@@ -6,16 +6,23 @@ pub struct SVGWrapper(Vec<u8>, String);
 
 impl SVGWrapper {
     pub fn evcxr_display(&self) {
-        let svg = String::from_utf8_lossy(self.0.as_slice());
-        println!(
-            "EVCXR_BEGIN_CONTENT text/html\n<div style=\"{}\">{}</div>\nEVCXR_END_CONTENT",
-            self.1, svg
-        );
+        println!("{:?}", self);
     }
 
     pub fn style<S: Into<String>>(mut self, style: S) -> Self {
         self.1 = style.into();
         self
+    }
+}
+
+impl std::fmt::Debug for SVGWrapper {
+    fn fmt(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+        let svg = String::from_utf8_lossy(self.0.as_slice());
+        write!(
+            formatter,
+            "EVCXR_BEGIN_CONTENT text/html\n<div style=\"{}\">{}</div>\nEVCXR_END_CONTENT",
+            self.1, svg
+        )
     }
 }