+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Imporvement
 
 - Histogram imporvements, horizental bar is supported, new creation API which compiler can infer the type
+- Supporting split the drawing area with a list of breakpoints using `DrawingArea::split_by_breakpoints`
 
 ### Fix